storcon: use wait LSN to kick the pageserver ingest
Problem

We've seen the ingest connection manager get stuck shortly after a
migration.

Summary of Changes

A speculative mitigation is to use the same mechanism as get page
requests for kicking LSN ingest. The connection manager monitors
LSN waits and queries the broker if no updates are received for the
timeline.