compute: Enable building VM images suitable for lauching with neonvm-guest
The 'legacy-vm-builder-image' Dockerfile target can still be used to
build an old-style VM image, for use with vm-builder. And that's still
the default target. However, there are now ano other target,
'compute-node-bootable', that can be used to build an image that
includes systemd unit files to launch the services, and
'compute-node-neonvm-payload' to package that as a QCOW2 image that
can be used with the new neon-vm launcher.

See https://github.com/neondatabase/autoscaling/pull/1211 for the new
launcher.