pageserver: hold GC while reading from a timeline
Problem

If we are GC-ing because a new image layer was added while traversing
the timeline, then it will remove layers that are required for fulfilling
the current get request (read-path cannot "look back" and notice the new
image layer).

Summary of Changes

Prevent GC from progressing on the current timeline while it is being
visited for a read.