tests: test_pgdata_import_smoke requires the 'testing' cargo feature
It took me ages to figure out why it was failing on my laptop. What I
saw was that when the test makes the 'import_pgdata' in the
pageserver, the pageserver actually performs a regular 'bootstrap'
timeline creation by running initdb, with no importing. It boiled down
to the json request that the test uses:

```
        {
            "new_timeline_id": str(timeline_id),
            "import_pgdata": {
                "idempotency_key": str(idempotency),
                "location": {"LocalFs": {"path": str(importbucket.absolute())}},
            },
        },
```

and how serde deserializes into rust structs. The 'LocalFs' enum
variant in `models.rs` is gated on the 'testing' cargo feature. On a
non-testing build, that got deserialized into the default Bootstrap
enum variant, as a valid TimelineCreateRequestModeImportPgdata variant
could not be formed.

PS. IMHO we should get rid of the testing feature, compile in all the
functionality, and have a runtime flag to disable anything
dangeorous. With that, you would've gotten a nice "feature only
enabled in testing mode" error in this case, or the test would've
simply worked. But that's another story.