storcon: silence cplane 404s on tenant creation
Problem

We get WARN log noise on tenant creations. Cplane creates tenants
via /location_config. That returns the attached locations in the
response and spawns a reconciliation which will also attempt to notify
cplane. If the notification is attempted before cplane persists the
shards to its database, storcon gets back a 404. The situation is
harmless, but annoying.

Summary of Changes

* Add a tenant creation hint to the reconciler config
* If the hint is true and we get back a 404 on the notification from
  cplane, ignore the error, but still queue the reconcile up for a
  retry.