safekeeper: fix the race by moving the current position back
The notification select arm runs when we want to attach a new reader
to the current cursor. It checks the current position of the cursor
and resets it if required.

The current position of the cursor is updated in the other select arm.
That runs when we get some WAL to send.

Now, assume we want to attach two shards consecutively to the cursor,
Let's say the notification select arm runs twice in a row. Say the cursor
is currently at LSN X. First shard wants to attach at position V and the
other one at W. Assume X > W > V.

First shard resets the stream to position V. Second shard comes in,
sees stale cursor position X and resets it to W. This means that the
first shard skips WAL in the [V, W) range.

Ultimately, this boils down to the current position not being kept in
sync with the reset of the WAL stream. This patch fixes the race by
updating it when resetting the WAL stream.