compute: Allow postgres user to power off the VM
I plan to use this when launching a fast_import job in a VM. There's
currently no good way for an executable running in a NeonVM to exit
gracefully and have the VM shut down. The inittab we use always
respawns the payload command. The idea is that the control plane can
use "fast_import ... && poweroff" as the command, so that when
fast_import completes successfully, the VM is terminated, and the k8s
Pod and VirtualMachine object are marked as completed successfully.

I'm working on bigger changes to how we launch VMs, and will try to
come up with a nicer system for that, but in the meanwhile, this quick
hack allows us to proceed with using VMs for one-off jobs like
fast_import.