Download awscli in separate layer in Dockerfile, to allow caching
The awscli was downloaded at the last stages of the overall compute
image build, which meant that if you modified any part of the build,
it would trigger a re-download of the awscli. That's a bit annoying
when developing locally and rebuilding the compute image
repeatedly. Move it to a separate layer, to cache separately and to
avoid the spurious rebuilds.