Link pgbouncer dynamically
I don't see the point of static linking, postgres itself and many of
the extensions are already built dynamically.

One reason for the change is that I'm working on bigger changes to
start using systemd in the compute, and as part of that I wanted to
add the --with-systemd configure option to pgbouncer, and there
doesn't seem to be a static version of libsystemd (at least not on
Debian).