fast_import: Make CPU & memory size configurable
The old values assumed that you have at least about 18 GB of RAM
available (shared_buffers=10GB and maintenance_work_mem=8GB).  That's
a lot when testing locally. Make it configurable, and make the default
assumption much smaller: 256 MB.

This is nice for local testing, but it's also in preparation for
starting to use VMs to run these jobs. When launched in a VM, the
control plane can set these env variables according to the max size of
the VM.

Also change the formula for how RAM is distributed: use 10% of RAM for
shared_buffers, and 70% for maintenance_work_mem. That leaves a good
amount for misc. other stuff and the OS. A very large shared_buffers
setting won't typically help with bulk loading. It won't help with the
network and I/O of processing all the tables, unless maybe if the
whole database fits in shared buffers, but even then it's not much
faster than using local disk. Bulk loading is all sequential I/O. It
also won't help much with index creation, which is also sequential
I/O.  A large maintenance_work_mem can be quite useful, however, so
that's where we put most of the RAM.