fast_import: Use rust s3 client for uploading
This replaces the use of the awscli utility. awscli binary is massive,
it added about 200 MB to the docker image size, while the s3 client
was already a dependency so using that is essentially free, as far as
binary size is concerned.

I implemented a simple upload function that tries to keep 10 uploads
going in parallel. I believe that's the default behavior of the "aws
s3 sync" command too.